Логика из компонента таблицы вынесена в хуки
