+++ components/daly-table/components/EditableRow.tsx
@@ -1,0 +1,5 @@
+export const EditableRow: React.FC<unknown> = (props) => {
+    return (
+        <tr {...props} style={{ cursor: 'pointer' }} />
+    );
+}
