+++ components/daly-table/daly-table.tsx
@@ -1,143 +1,40 @@
-import React, { useEffect, useState } from 'react';
-import { Table } from 'antd';
-import { dalyTableSelectors, dalyTableThunks, IDalyTableItemTask } from './daly-table-slice';
-import { useSelector } from 'react-redux';
-import { useAppDispatch } from '../../redux';
-import { CreateButton } from './components/create-button';
-import { format, startOfDay } from 'date-fns';
+import React, { useState } from 'react';
+import { Button, Table } from 'antd';
 import { CreateEditModalForm } from './components/create-edit-modal-form';
-import { fetchErrors } from '../../common-types';
-
-const EditableRow: React.FC<unknown> = (props) => {
-    return (
-        <tr {...props} style={{ cursor: 'pointer' }} />
-    );
-}
+import { useDalyTableColumns } from './use-daly-table-columns';
+import { useDalyTableDataSource } from './use-daly-table-data-source';
+import { useDalyTableEditItemData } from './use-daly-table-edit-item-data';
 
 export const DalyTable: React.FC = () => {
-    const dataSource = useSelector(dalyTableSelectors.selectAll);
-    const [visibleEditForm, setVisibleEditForm] = useState(false);
-    const [editFormData, setEditFormData] = useState<{
-        timeStart: string,
-        timeEnd: string,
-        keyTask: number,
-        desctiption: string
-    }>(null);
-    const [keyEditItem, setKeyEditItem] = useState<number>();
-    const dispatch = useAppDispatch();
-
-    useEffect(() => {
-        dispatch(dalyTableThunks.fetchDalyItems());
-    }, []);
-
-    const renderTimeFromUtc = (dateStrUtc: string) => format(new Date(dateStrUtc), 'HH:mm:ss');
-
-    const columnsForTable = [
-        {
-            title: 'Дата',
-            dataIndex: 'date',
-            render: (dateStrUtc: string) => format(new Date(dateStrUtc), 'dd.MM.yyyy')
-        },
-        {
-            title: 'Номер таска',
-            dataIndex: 'keyTask'
-        },
-        {
-            title: 'Время начала',
-            dataIndex: 'timeStart',
-            render: renderTimeFromUtc
-        },
-        {
-            title: 'Время конца',
-            dataIndex: 'timeEnd',
-            render: renderTimeFromUtc
-        },
-        {
-            title: 'Описание',
-            dataIndex: 'desctiption'
-        },
-        {
-            title: 'Действия',
-            dataIndex: 'actions',
-            render: (_text: any, record: IDalyTableItemTask) => {
-                
-                return 'В разработке :)';
-            }
-        }
-    ];
-
-    const components = {
-        body: {
-            row: EditableRow
-        },
-    };
+    const [visibleCreateEditForm, setVisibleCreateEditForm] = useState(false);
+    const dataSource = useDalyTableDataSource();
+    const columns = useDalyTableColumns();
+    const {
+        tableProps,
+        createEditModalFormProps
+    } = useDalyTableEditItemData(setVisibleCreateEditForm);
 
     return (
         <>
-            {/* <CreateButton
-                initialValues={{
-                    date: startOfDay(new Date()).toUTCString(),
-                    keyTask: dataSource[dataSource.length - 1]?.keyTask,
-                    timeStart: dataSource[dataSource.length - 1]?.timeEnd,
-                    timeEnd: new Date().toUTCString()
-                }}
-            /> */}
+            <Button 
+                type='primary' 
+                 
+                style={{ marginBottom: 16, alignSelf: 'flex-start' }}
+            >
+                Добавить запись (Ctrl+Shift+a)
+            </Button>
             <Table
-                components={components}
                 rowClassName={() => 'editable-row'}
                 className='daly-table'
                 bordered
                 dataSource={dataSource}
-                columns={columnsForTable}
-                onRow={(record) => ({
-                    onClick() {
-                        setEditFormData({
-                            timeStart: record.timeStart,
-                            timeEnd: record.timeEnd,
-                            desctiption: record.desctiption,
-                            keyTask: record.keyTask
-                        });
-                        setKeyEditItem(record.key);
-                        setVisibleEditForm(true);
-                    }
-                })}
+                columns={columns}
+                {...tableProps}
             />
             <CreateEditModalForm
-                onCancel={() => {
-                    setEditFormData(null);
-                    setKeyEditItem(null);
-                    setVisibleEditForm(false);
-                }}
-                onCreate={(newDalyItem) => {
-                    return new Promise<0 | 1>(async (resolve, reject) => {
-                        if(Object.getOwnPropertyNames(newDalyItem).every(
-                            key => newDalyItem[key] === editFormData[key]
-                        )) resolve(1);
-                        else {
-                            const value = await dispatch(dalyTableThunks.fetchEditItem({
-                                key: keyEditItem,
-                                timeStart: newDalyItem.timeStart,
-                                timeEnd: newDalyItem.timeEnd,
-                                desctiption: newDalyItem.desctiption,
-                                keyTask: newDalyItem.keyTask
-                            }));
-    
-                            if (value.payload === fetchErrors.common) reject();
-                            else {
-                                setVisibleEditForm(false);
-                                resolve(0);
-                            }
-                        }
-                    })
-                }}
-                visible={visibleEditForm}
-                initialValues={{
-                    desctiption: editFormData?.desctiption,
-                    keyTask: editFormData?.keyTask,
-                    timeEnd: editFormData?.timeEnd,
-                    timeStart: editFormData?.timeStart
-                }}
+                visible={visibleCreateEditForm}
                 isEdit
+                {...createEditModalFormProps}
             />
         </>
     )
